Move HTTP related code to own submodule
